Revert chat options theme move
