+++ packages/ag-charts-types/src/chart/chartOptions.ts
@@ -11,7 +11,6 @@ import type { AgChartLegendOptions } from './legendOptions';
 import type { AgLocaleOptions } from './localeOptions';
 import type { AgNavigatorOptions } from './navigatorOptions';
 import type { AgRangesOptions } from './rangesOptions';
-import type { AgChartTheme, AgChartThemeName } from './themeOptions';
 import type { AgChartTooltipOptions } from './tooltipOptions';
 import type { CssColor, FontFamily, FontSize, FontStyle, FontWeight, PixelSize, TextAlign, TextWrap } from './types';
 import type { AgZoomOptions } from './zoomOptions';
@@ -252,10 +251,4 @@ export interface AgBaseChartOptions<TDatum = any> extends AgBaseThemeableChartOp
     container?: HTMLElement | null;
     /** The initial state of the chart. This must be a serializable value. */
     initialState?: AgInitialStateOptions;
-    /**
-     * A predefined theme name or an object containing theme overrides.
-     *
-     * See: [Themes Reference](/themes-api/)
-     */
-    theme?: AgChartTheme | AgChartThemeName;
 }
