+++ packages/ag-charts-types/src/chartBuilderOptions.ts
@@ -1,7 +1,12 @@
 /* eslint-disable sonarjs/class-name */
 import type { AgInitialStateOptions } from './api/initialStateOptions';
 import type { AgBaseChartOptions } from './chart/chartOptions';
-import type { AgBaseChartThemeOptions, AgBaseGaugePresetThemeOptions } from './chart/themableOptions';
+import type {
+    AgBaseChartThemeOptions,
+    AgBaseGaugePresetThemeOptions,
+    AgChartTheme,
+    AgChartThemeName,
+} from './chart/themeOptions';
 import type { PixelSize } from './chart/types';
 import type { AgFinancialChartPresets } from './presets/financial/financialOptions';
 import type { AgGaugePresets } from './presets/gauge/gaugeOptions';
@@ -23,13 +28,33 @@ import type { AgBaseTopologyChartOptions } from './series/topology/topologyOptio
 
 export interface AgChartThemeOptions extends AgBaseChartThemeOptions {}
 
-export interface AgCartesianChartOptions extends AgBaseCartesianChartOptions, AgBaseChartOptions {}
-export interface AgPolarChartOptions extends AgBasePolarChartOptions, AgBaseChartOptions {}
-export interface AgHierarchyChartOptions extends AgBaseHierarchyChartOptions, AgBaseChartOptions {}
-export interface AgTopologyChartOptions extends AgBaseTopologyChartOptions, AgBaseChartOptions {}
-export interface AgFlowProportionChartOptions extends AgBaseFlowProportionChartOptions, AgBaseChartOptions {}
-export interface AgStandaloneChartOptions extends AgBaseStandaloneChartOptions, AgBaseChartOptions {}
-export interface AgGaugeChartOptions extends AgBaseChartOptions {}
+export interface AgCartesianChartOptions extends AgBaseCartesianChartOptions, AgBaseChartOptions {
+    /**
+     * A predefined theme name or an object containing theme overrides.
+     *
+     *
+     * See: [Themes Reference](/themes-api/)
+     */
+    theme?: AgChartTheme | AgChartThemeName;
+}
+export interface AgPolarChartOptions extends AgBasePolarChartOptions, AgBaseChartOptions {
+    theme?: AgChartTheme | AgChartThemeName;
+}
+export interface AgHierarchyChartOptions extends AgBaseHierarchyChartOptions, AgBaseChartOptions {
+    theme?: AgChartTheme | AgChartThemeName;
+}
+export interface AgTopologyChartOptions extends AgBaseTopologyChartOptions, AgBaseChartOptions {
+    theme?: AgChartTheme | AgChartThemeName;
+}
+export interface AgFlowProportionChartOptions extends AgBaseFlowProportionChartOptions, AgBaseChartOptions {
+    theme?: AgChartTheme | AgChartThemeName;
+}
+export interface AgStandaloneChartOptions extends AgBaseStandaloneChartOptions, AgBaseChartOptions {
+    theme?: AgChartTheme | AgChartThemeName;
+}
+export interface AgGaugeChartOptions extends AgBaseChartOptions {
+    theme?: AgChartTheme | AgChartThemeName;
+}
 export type AgChartOptions =
     | AgCartesianChartOptions
     | AgPolarChartOptions
@@ -65,18 +90,21 @@ export type AgBaseSparklinePresetThemeOptions = AgSparklineBaseThemeableOptions 
         | 'padding'
         | 'width'
         | 'data'
-        | 'theme'
     >;
 
 export type AgFinancialChartOptions = AgBaseFinancialPresetOptions & AgFinancialChartPresets;
 
-export interface AgBaseGaugePresetOptions extends AgBaseGaugePresetThemeOptions {}
+export interface AgBaseGaugePresetOptions extends AgBaseGaugePresetThemeOptions {
+    theme?: AgChartTheme | AgChartThemeName;
+}
 
 export type AgLinearGaugeOptions = AgBaseGaugePresetOptions & AgLinearGaugePreset;
 export type AgRadialGaugeOptions = AgBaseGaugePresetOptions & AgRadialGaugePreset;
 export type AgGaugeOptions = AgBaseGaugePresetOptions & AgGaugePresets;
 
-export interface AgBaseSparklinePresetOptions extends AgBaseSparklinePresetThemeOptions {}
+export interface AgBaseSparklinePresetOptions extends AgBaseSparklinePresetThemeOptions {
+    theme?: AgChartTheme | AgChartThemeName;
+}
 
 export type AgBarSparklineOptions = AgBaseSparklinePresetOptions & AgBarSparklinePreset;
 export type AgLineSparklineOptions = AgBaseSparklinePresetOptions & AgLineSparklinePreset;
