+++ packages/ag-charts-types/src/main.ts
@@ -24,7 +24,6 @@ export * from './chart/polarAxisOptions';
 export * from './chart/radiusAxisOptions';
 export * from './chart/rangesOptions';
 export * from './chart/themeOptions';
-export * from './chart/themableOptions';
 export * from './chart/themeParamsOptions';
 export * from './chart/tooltipOptions';
 export * from './chart/types';
